sin-trade-be/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==5.0.0
gunicorn==23.0.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pluggy==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
Werkzeug==3.0.3
supabase==2.10.0
pika==1.3.2
apscheduler==3.11.0
requests==2.32.5
resend==2.23.0